Restores conceal setting to 0 to address json pain
The indentLine plugin sets the conceal value to 2 and the impact of this
is that when we work with json and markdown files, some characters e.g.
quotes (") are "hidden" in edit mode but present in visual mode. This is
okay, but god is it pain when  you need to work with json and we can't
see the characters being edited.